Fix idiotic error in movie state load code - code was reading description from state as DWORDs and then saving back only the bottom byte, also messing up the movie data itself

+++ src/Movie.cpp
@@ -810,7 +810,7 @@ int	Load (FILE *in)
 		tpi = tpl;					// INFO
 		while (tpi > 0)
 		{
-			readLong(tpc);
+			readByte(tpc);
 			fwrite(&tpc, 1, 1, Data);
 			tpi--;
 		}
